# docs/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: IoC Cognition Fabric Node service — shared memory routing and memory
    operations proxy.
  title: CFN Service API
  version: "1.0"
paths:
...
  /api/workspaces/{workspaceId}/multi-agentic-systems/{masId}/agents/{agentId}/memory-operations:
    post:
      consumes:
      - application/json
      description: |-
        Forwards REST API requests to a remote memory provider (Mem0, Graphiti, etc.) for agent-specific memory operations.
        The memory provider base URL and auth credentials are auto-resolved from management plane config based on workspace/MAS/agent IDs.
        The `http-url` field should contain the relative path and query parameters to append to the provider base URL.

        **GET example** — retrieve memories:
        ```json
        {
        "header": {},
        "payload": {
        "http-request-type": "GET",
        "http-url": "v1/memories/?user_id=curl-test-user",
        "http-request-body": {},
        "http-headers": {}
        }
        }
        ```

        **POST example** — add memories:
        ```json
        {
        "header": {},
        "payload": {
        "http-request-type": "POST",
        "http-url": "/v1/memories/",
        "http-request-body": {
        "messages": [{"role": "user", "content": "I prefer dark mode in all my apps"}],
        "user_id": "curl-test-user"
        },
        "http-headers": {}
        }
        }
        ```
      parameters:
      - description: Workspace ID
        in: path
        name: workspaceId
        required: true
        type: string
      - description: Multi-Agentic System ID
        in: path
        name: masId
        required: true
        type: string
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      - description: Memory operation request (see MemoryOperationRequest)
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Proxied response (actual provider status is in http-status
            field)
          schema:
            type: object
        "400":
          description: Invalid request body or missing http-request-type
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Memory provider config not found for agent
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Failed to forward request to memory provider
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Memory proxy client not configured
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Proxy API requests to a remote memory provider
      tags:
      - memory-operations
...
swagger: "2.0"